chore: address review comment

+++ ddtrace/internal/settings/profiling.pyi
@@ -29,7 +29,7 @@ class ProfilingConfigStack(DDConfig):
     enabled: bool
     adaptive_sampling: bool
     adaptive_sampling_target_overhead: float
-    adaptive_sampling_max_interval: float
+    adaptive_sampling_max_interval: int
     native_frames: bool
 
 class ProfilingConfigLock(DDConfig):
